Move tests to unit test project
