+++ DfE.FindInformationAcademiesTrusts.sln
@@ -25,8 +25,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DfE.FindInformationAcademie
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DfE.FindInformationAcademiesTrusts.Data.FiatDb.UnitTests", "tests\DfE.FindInformationAcademiesTrusts.Data.FiatDb.UnitTests\DfE.FindInformationAcademiesTrusts.Data.FiatDb.UnitTests.csproj", "{187B33C1-628B-4FF4-9A7E-C775F23F6699}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DfE.FindInformationAcademiesTrusts.IntegrationTests", "tests\DfE.FindInformationAcademiesTrusts.IntegrationTests\DfE.FindInformationAcademiesTrusts.IntegrationTests.csproj", "{73980026-EB6D-4980-9A96-E30ADADFB3B2}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -73,10 +71,6 @@ Global
 		{187B33C1-628B-4FF4-9A7E-C775F23F6699}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{187B33C1-628B-4FF4-9A7E-C775F23F6699}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{187B33C1-628B-4FF4-9A7E-C775F23F6699}.Release|Any CPU.Build.0 = Release|Any CPU
-		{73980026-EB6D-4980-9A96-E30ADADFB3B2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{73980026-EB6D-4980-9A96-E30ADADFB3B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{73980026-EB6D-4980-9A96-E30ADADFB3B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{73980026-EB6D-4980-9A96-E30ADADFB3B2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -87,7 +81,6 @@ Global
 		{90AA9D53-B5B0-489F-AEC1-23551EA04BC5} = {E22DC2DA-4EB5-41C6-A41D-349692BCCDBE}
 		{1DF57DAB-F9BA-4810-BD35-BB59996FF7BB} = {E22DC2DA-4EB5-41C6-A41D-349692BCCDBE}
 		{187B33C1-628B-4FF4-9A7E-C775F23F6699} = {E22DC2DA-4EB5-41C6-A41D-349692BCCDBE}
-		{73980026-EB6D-4980-9A96-E30ADADFB3B2} = {E22DC2DA-4EB5-41C6-A41D-349692BCCDBE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CEE3F8EE-1688-4C85-8749-E2B3FCB9A5FE}
