--- tests/DfE.FindInformationAcademiesTrusts.IntegrationTests/Base/BaseIntegrationTest.cs
@@ -1,6 +1,0 @@
-namespace DfE.FindInformationAcademiesTrusts.IntegrationTests.Base;
-
-[Collection(nameof(UseApplicationWithMockedServices))]
-public abstract class BaseIntegrationTest
-{
-}
