+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/WebHost/AuthorisedUserTests.cs
@@ -1,32 +1,30 @@
 using System.Net;
-using DfE.FindInformationAcademiesTrusts.IntegrationTests.AuthenticationHandlers;
-using DfE.FindInformationAcademiesTrusts.IntegrationTests.Base;
-using FluentAssertions;
+using DfE.FindInformationAcademiesTrusts.UnitTests.WebHost.AuthenticationHandlers;
+using DfE.FindInformationAcademiesTrusts.UnitTests.WebHost.Base;
 using FluentAssertions.Execution;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
 
-namespace DfE.FindInformationAcademiesTrusts.IntegrationTests;
+namespace DfE.FindInformationAcademiesTrusts.UnitTests.WebHost;
 
 /// <summary>
-/// Tests for when the user is authenticated but doesn't have the correct role
+/// Tests for when the user is authenticated and has the authorised role
 /// </summary>
-public class UnauthorisedUserTests : BaseIntegrationTest
+public class AuthorisedUserTests : BaseWebHostTest
 {
     private readonly HttpClient _client;
 
-    public UnauthorisedUserTests(ApplicationWithMockedServices factory)
+    public AuthorisedUserTests(ApplicationWithMockedServices factory)
     {
         _client = factory.WithWebHostBuilder(builder =>
             {
                 builder.ConfigureTestServices(services =>
                 {
-                    //Override authentication to simulate unauthorised user
-                    services.AddAuthentication(UnauthorisedUserAuthenticationHandler.AuthenticationScheme)
-                        .AddScheme<AuthenticationSchemeOptions, UnauthorisedUserAuthenticationHandler>(
-                            UnauthorisedUserAuthenticationHandler.AuthenticationScheme,
+                    services.AddAuthentication(AuthorisedUserAuthenticationHandler.AuthenticationScheme)
+                        .AddScheme<AuthenticationSchemeOptions, AuthorisedUserAuthenticationHandler>(
+                            AuthorisedUserAuthenticationHandler.AuthenticationScheme,
                             options => { options.ForwardForbid = CookieAuthenticationDefaults.AuthenticationScheme; }
                         );
                 });
@@ -38,8 +36,9 @@ public class UnauthorisedUserTests : BaseIntegrationTest
     }
 
     [Theory]
+    [MemberData(nameof(FiatPages.ExpectedProtectedRoutesInAppMemberData), MemberType = typeof(FiatPages))]
     [MemberData(nameof(FiatPages.ExpectedAlwaysAccessibleRoutesMemberData), MemberType = typeof(FiatPages))]
-    public async Task Unprotected_page_renders_when_unauthorised(string url)
+    public async Task Protected_page_renders_when_authorised(string url)
     {
         var response = await _client.GetAsync(url);
 
@@ -64,20 +63,29 @@ public class UnauthorisedUserTests : BaseIntegrationTest
         pageContent.Should().Contain("Healthy");
     }
 
-    [Theory]
-    [MemberData(nameof(FiatPages.ExpectedProtectedRoutesInAppMemberData), MemberType = typeof(FiatPages))]
-    [InlineData("notfound")]
-    public async Task Protected_pages_redirect_to_no_access_when_unauthorised(string url)
+    [Fact]
+    public async Task Not_found_page_renders_when_authorised()
     {
-        var response = await _client.GetAsync(url);
+        var response = await _client.GetAsync("/notfound");
+
+        using var _ = new AssertionScope();
+
+        response.StatusCode.Should().Be(HttpStatusCode.NotFound);
+
+        var pageContent = await response.Content.ReadAsStringAsync();
+        pageContent.Should().Contain("Find information about academies and trusts");
+    }
+
+    [Fact]
+    public async Task No_access_redirects_to_return_url()
+    {
+        var response = await _client.GetAsync($"no-access?ReturnUrl={WebUtility.UrlEncode("/search")}");
 
         using var _ = new AssertionScope();
 
         response.StatusCode.Should().Be(HttpStatusCode.Redirect);
 
         response.Headers.Location.Should().NotBeNull();
-        response.Headers.Location!.Host.Should().Be("localhost");
-        response.Headers.Location!.Segments.Should().Contain("no-access");
-        response.Headers.Location!.Query.Should().Be($"?ReturnUrl={WebUtility.UrlEncode($"/{url}")}");
+        response.Headers.Location!.OriginalString.Should().Be("/search");
     }
 }
