+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/WebHost/Base/BaseWebHostTest.cs
@@ -1,0 +1,6 @@
+namespace DfE.FindInformationAcademiesTrusts.UnitTests.WebHost.Base;
+
+[Collection(nameof(UseApplicationWithMockedServices))]
+public abstract class BaseWebHostTest
+{
+}
