+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/WebHost/UnauthorisedUserTests.cs
@@ -1,31 +1,31 @@
 using System.Net;
-using DfE.FindInformationAcademiesTrusts.IntegrationTests.AuthenticationHandlers;
-using DfE.FindInformationAcademiesTrusts.IntegrationTests.Base;
-using FluentAssertions;
+using DfE.FindInformationAcademiesTrusts.UnitTests.WebHost.AuthenticationHandlers;
+using DfE.FindInformationAcademiesTrusts.UnitTests.WebHost.Base;
 using FluentAssertions.Execution;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
 
-namespace DfE.FindInformationAcademiesTrusts.IntegrationTests;
+namespace DfE.FindInformationAcademiesTrusts.UnitTests.WebHost;
 
 /// <summary>
-/// Tests for when the user is authenticated and has the authorised role
+/// Tests for when the user is authenticated but doesn't have the correct role
 /// </summary>
-public class AuthorisedUserTests : BaseIntegrationTest
+public class UnauthorisedUserTests : BaseWebHostTest
 {
     private readonly HttpClient _client;
 
-    public AuthorisedUserTests(ApplicationWithMockedServices factory)
+    public UnauthorisedUserTests(ApplicationWithMockedServices factory)
     {
         _client = factory.WithWebHostBuilder(builder =>
             {
                 builder.ConfigureTestServices(services =>
                 {
-                    services.AddAuthentication(AuthorisedUserAuthenticationHandler.AuthenticationScheme)
-                        .AddScheme<AuthenticationSchemeOptions, AuthorisedUserAuthenticationHandler>(
-                            AuthorisedUserAuthenticationHandler.AuthenticationScheme,
+                    //Override authentication to simulate unauthorised user
+                    services.AddAuthentication(UnauthorisedUserAuthenticationHandler.AuthenticationScheme)
+                        .AddScheme<AuthenticationSchemeOptions, UnauthorisedUserAuthenticationHandler>(
+                            UnauthorisedUserAuthenticationHandler.AuthenticationScheme,
                             options => { options.ForwardForbid = CookieAuthenticationDefaults.AuthenticationScheme; }
                         );
                 });
@@ -37,9 +37,8 @@ public class AuthorisedUserTests : BaseIntegrationTest
     }
 
     [Theory]
-    [MemberData(nameof(FiatPages.ExpectedProtectedRoutesInAppMemberData), MemberType = typeof(FiatPages))]
     [MemberData(nameof(FiatPages.ExpectedAlwaysAccessibleRoutesMemberData), MemberType = typeof(FiatPages))]
-    public async Task Protected_page_renders_when_authorised(string url)
+    public async Task Unprotected_page_renders_when_unauthorised(string url)
     {
         var response = await _client.GetAsync(url);
 
@@ -64,29 +63,20 @@ public class AuthorisedUserTests : BaseIntegrationTest
         pageContent.Should().Contain("Healthy");
     }
 
-    [Fact]
-    public async Task Not_found_page_renders_when_authorised()
+    [Theory]
+    [MemberData(nameof(FiatPages.ExpectedProtectedRoutesInAppMemberData), MemberType = typeof(FiatPages))]
+    [InlineData("notfound")]
+    public async Task Protected_pages_redirect_to_no_access_when_unauthorised(string url)
     {
-        var response = await _client.GetAsync("/notfound");
-
-        using var _ = new AssertionScope();
-
-        response.StatusCode.Should().Be(HttpStatusCode.NotFound);
-
-        var pageContent = await response.Content.ReadAsStringAsync();
-        pageContent.Should().Contain("Find information about academies and trusts");
-    }
-
-    [Fact]
-    public async Task No_access_redirects_to_return_url()
-    {
-        var response = await _client.GetAsync($"no-access?ReturnUrl={WebUtility.UrlEncode("/search")}");
+        var response = await _client.GetAsync(url);
 
         using var _ = new AssertionScope();
 
         response.StatusCode.Should().Be(HttpStatusCode.Redirect);
 
         response.Headers.Location.Should().NotBeNull();
-        response.Headers.Location!.OriginalString.Should().Be("/search");
+        response.Headers.Location!.Host.Should().Be("localhost");
+        response.Headers.Location!.Segments.Should().Contain("no-access");
+        response.Headers.Location!.Query.Should().Be($"?ReturnUrl={WebUtility.UrlEncode($"/{url}")}");
     }
 }
